Prefactor: move test utils package
